// src/app/definitions.d.ts
export type UserId = string;
export type RoomId = string;
export type GameId = string;

export interface DatabaseGame {
  gameId: GameId;
  gameParameters: GameParameters;
  firstPlayerId: UserId;
  secondPlayerId: UserId;
  gameDate: EpochTimeStamp;
  status: GameStatus;
  moves?: string;
}

export interface Game {
  game_id: string;
  game_parameters: GameParameters | LocalGameParameters;
  grid: Array<Array<number>>;
  first_player_id: string;
  second_player_id: string;
  turn: number;
}

export interface GameParameters {
  ranked: boolean;
  board_size: number;
  time_limit: number;
}
export interface LocalGameParameters {
  time_limit: number,
  board_size: number
}